requirements.txt
discord.py==2.6.4
aiohttp==3.13.1
holidays==0.83
beautifulsoup4~=4.14.2
schedule~=1.2.2
python-dotenv~=1.1.1
tzdata==2025.2
httpx
lxml
